+++ kernel/drivers/fbdev.h
@@ -16,7 +16,7 @@ typedef struct {
     uint8_t blue_mask_shift;
 } __attribute__((packed)) framebuffer_info_t;
 
-int fbdev_read(int minor_number, uint64_t offset, uint8_t *buffer, uint64_t size);
-int fbdev_write(int minor_number, uint64_t offset, const uint8_t *buffer, uint64_t size);
+int64_t fbdev_read(int minor_number, uint64_t offset, uint8_t *buffer, uint64_t size);
+int64_t fbdev_write(int minor_number, uint64_t offset, const uint8_t *buffer, uint64_t size);
 int fbdev_ioctl(int minor_number, uint64_t request, uint64_t arg);
 void fbdev_init(volatile struct limine_framebuffer_request* req);
