+++ libc/fd.c
@@ -2,7 +2,7 @@
 #include "errno.h"
 #include "unistd.h"
 
-int read(int path, void* buffer, size_t size) {
+int64_t read(int path, void* buffer, size_t size) {
     int res = syscall(SYSCALL_READ, (uint64_t)path, (uint64_t)buffer, (uint64_t)size, 0, 0, 0);
     if (res < 0) {
         errno = -res;
@@ -11,7 +11,7 @@ int read(int path, void* buffer, size_t size) {
     return res;
 }
 
-int write(int path, const void* buffer, size_t size) {
+int64_t write(int path, const void* buffer, size_t size) {
     int res = syscall(SYSCALL_WRITE, (uint64_t)path, (uint64_t)buffer, (uint64_t)size, 0, 0, 0);
     if (res < 0) {
         errno = -res;
@@ -38,7 +38,7 @@ int close(int fd) {
     return res;
 }
 
-int seek(int fd, int64_t offset, int type) {
+int64_t seek(int fd, int64_t offset, int type) {
     int res = syscall(SYSCALL_SEEK, (uint64_t)fd, (uint64_t)offset, (uint64_t)type, 0, 0, 0);
     if (res < 0) {
         errno = -res;
